agent-ui: refactor code
