+++ packages/agent-ui/src/actions/actionTypes.js
@@ -1,0 +1,2 @@
+export const AGENT_INFO_SUCCESS = 'AGENT_INFO_SUCCESS';
+export const AGENT_INFO_FAILURE = 'AGENT_INFO_FAILURE';
