+++ packages/agent-ui/src/actions/getAgentInfo.js
@@ -1,13 +1,9 @@
 import { getAgent } from 'stratumn-agent-client';
+import { getAgentInfoSuccess } from './';
 
 const url = 'http://localhost:3000/';
 
-export function getAgentInfoSuccess(info) {
-  console.log('getAgentInfoSuccess', info);
-  return { type: 'AGENT_INFO_SUCCESS', info };
-}
-
-export function getAgentInfo() {
+export default function getAgentInfo() {
   return function withDispatch(dispatch) {
     return getAgent(url)
       .then(res => dispatch(getAgentInfoSuccess(res)))
