+++ packages/agent-ui/src/actions/getAgentInfoSuccess.js
@@ -1,0 +1,6 @@
+import { actionTypes } from './';
+
+export default function getAgentInfoSuccess(info) {
+  console.log('getAgentInfoSuccess', info);
+  return { type: actionTypes.AGENT_INFO_SUCCESS, info };
+}
