+++ packages/agent-ui/src/actions/index.js
@@ -1,0 +1,6 @@
+import * as _actionTypes from './actionTypes';
+
+export const actionTypes = _actionTypes;
+
+export { default as getAgentInfo } from './getAgentInfo';
+export { default as getAgentInfoSuccess } from './getAgentInfoSuccess';
