+++ packages/agent-ui/src/components/index.js
@@ -1,0 +1,3 @@
+export { default as App } from './app';
+export { default as LeftDrawer } from './leftDrawer';
+export { default as TopBar } from './topBar';
