+++ packages/agent-ui/src/index.js
@@ -1,21 +1,3 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
+import { createReactDOM } from './store';
 
-import { Provider } from 'react-redux';
-import { Router, browserHistory } from 'react-router';
-import { syncHistoryWithStore } from 'react-router-redux';
-
-import routes from './routes';
-import configureStore from './store/configure';
-
-const store = configureStore();
-
-// Create an enhanced history that syncs navigation events with the store
-const history = syncHistoryWithStore(browserHistory, store);
-
-ReactDOM.render(
-  <Provider store={store}>
-    <Router history={history} routes={routes} />
-  </Provider>,
-  document.getElementById('root')
-);
+createReactDOM(document.getElementById('root'));
