+++ packages/agent-ui/src/reducers/getAgentInfo.js
@@ -1,7 +1,9 @@
+import { actionTypes } from '../actions';
+
 export default function(state = null, action) {
   console.log('getAgentInfoReducer');
   switch (action.type) {
-    case 'AGENT_INFO_SUCCESS':
+    case actionTypes.AGENT_INFO_SUCCESS:
       return action.info;
     default:
       return state;
