+++ packages/agent-ui/src/reducers/index.js
@@ -1,0 +1,1 @@
+export { default as getAgentInfo } from './getAgentInfo';
