+++ packages/agent-ui/src/store/dom.js
@@ -1,0 +1,20 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+import { Provider } from 'react-redux';
+import { Router, browserHistory } from 'react-router';
+import { syncHistoryWithStore } from 'react-router-redux';
+
+import routes from '../routes';
+import { configureStore } from './';
+
+export default function(doc) {
+  const store = configureStore();
+  const history = syncHistoryWithStore(browserHistory, store);
+  ReactDOM.render(
+    <Provider store={store}>
+      <Router history={history} routes={routes} />
+    </Provider>,
+    doc
+  );
+}
