+++ packages/agent-ui/src/store/index.js
@@ -1,0 +1,2 @@
+export { default as createReactDOM } from './dom';
+export { default as configureStore } from './configure';
